raport-creator/requirements.txt
clickhouse-connect
fastapi
uvicorn
pandas
matplotlib
seaborn
python-docx
python-dotenv
jira